Updated EventRunner.
- Switched event storage from ArrayLists to an experimental Hashtable system.

This should allow for improved speed and stability, as well as reducing the amount of hard-coded events. Rather than being forced to insert events at specific points, it is possible to insert them just about anywhere along the cycle and it won't matter.

As an added bonus, the EventRunner is also easier to use now, since you don't need to input very specific values when adding events.